auth/requirements.txt
flask==3.1.2
flask-cors==6.0.1

gunicorn==23.0.0

PyJWT==2.10.1
